backend/requirements.txt
discord.py-self
python-dotenv
Pillow
Flask
Werkzeug
gunicorn
